requirements.txt
certifi==2020.11.8
chardet
click
Flask
idna
itsdangerous
Jinja2
MarkupSafe
python-dotenv
requests
urllib3
Werkzeug
